+++ docs/api_static/bapsflib._hdf.maps.controls.phys180e.rst
@@ -1,11 +1,11 @@
 :orphan:
 
-bapsflib\.\_hdf\.maps\.controls\.phys180e
-=========================================
+bapsflib\.\_hdf\.maps\.controls\.positions180e
+==============================================
 
-.. py:currentmodule:: bapsflib._hdf.maps.controls.phys180e
+.. py:currentmodule:: bapsflib._hdf.maps.controls.positions180e
 
-.. automodapi:: bapsflib._hdf.maps.controls.phys180e
+.. automodapi:: bapsflib._hdf.maps.controls.positions180e
     :no-heading:
     :include-all-objects:
     :inherited-members:
